services/main-app/requirements.txt
# FastAPI and server
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# Pydantic and validation
pydantic==2.5.3
pydantic-settings==2.1.0
email-validator==2.1.0

# Authentication and security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.1.2

# LangChain and AI
langchain==0.1.0
langchain-community==0.0.10
langchain-openai==0.0.2
openai==1.6.1
tiktoken==0.5.2

# Vector stores
chromadb==0.4.22
# weaviate-client==4.4.1
# pinecone-client==3.0.0
# qdrant-client==1.7.0

# Embeddings and reranking
cohere==4.39.0
sentence-transformers==2.2.2

# Document processing
pypdf==3.17.4
python-docx==1.1.0
openpyxl==3.1.2
python-pptx==0.6.23
pytesseract==0.3.10
Pillow==10.2.0
markdown==3.5.1
beautifulsoup4==4.12.2
chardet==5.2.0

# Database
# sqlalchemy==2.0.25
# asyncpg==0.29.0
# alembic==1.13.1

# Redis for caching
redis==5.0.1
# aioredis==2.0.1

# HTTP clients
httpx==0.26.0
aiohttp==3.9.1

# Utilities
python-dotenv==1.0.0
pyyaml==6.0.1
orjson==3.9.10

# Monitoring and logging
prometheus-client==0.19.0
structlog==24.1.0

# Testing (optional, for development)
pytest==7.4.4
pytest-asyncio==0.23.3
pytest-cov==4.1.0
httpx==0.26.0  # For testing

# AWS SDK (for S3 storage)
boto3==1.34.14

# Rate limiting and caching
slowapi==0.1.9
cachetools==5.3.2